+++ CHANGES.md
@@ -39,6 +39,7 @@
 - Fix problem with multiple exports [Taiga #2468](https://tree.taiga.io/project/penpot/issue/2468)
 - Allow import to continue from recoverable failures [#1412](https://github.com/penpot/penpot/issues/1412)
 - Improved behaviour on text options when not text is selected [Taiga #2390](https://tree.taiga.io/project/penpot/issue/2390)
+- Fix decimal numbers in export viewbox [Taiga #2290](https://tree.taiga.io/project/penpot/issue/2290)
 
 ### :arrow_up: Deps updates
 ### :heart: Community contributions by (Thank you!)